Added actual balance, updatating the money and fixed some bugs

+++ src/views/currency_management.py
@@ -1,7 +1,12 @@
+from typing import Any
+
 import discord
+from discord.ext.commands import NotOwner
+from discord.ui import Item
 
 from services.config import Config
 from services.currency import Currency
+from utils import errors
 from utils.context import Context
 
 
@@ -14,6 +19,11 @@ class ManageCurrencyView(discord.ui.View):
     @property
     def currency(self) -> Currency:
         return self._currency
+
+    async def on_error(
+        self, interaction: discord.Interaction, error: Exception, _: Item[Any]
+    ) -> None:
+        await errors.global_error_handler(interaction, error)
 
 
 class AddCurrencyView(ManageCurrencyView):
@@ -30,8 +40,10 @@ class AddCurrencyView(ManageCurrencyView):
             await interaction.message.delete()
 
     @discord.ui.button(label="No", style=discord.ButtonStyle.red)
-    @Config.has_permission("manage_currencies")
     async def no(self, interaction: discord.Interaction, _: discord.ui.Button) -> None:
+        if interaction.user.id != self._ctx.author.id:
+            raise NotOwner("You do not own this currency")
+
         if interaction.message:
             await interaction.message.delete()
 
@@ -41,10 +53,12 @@ class DeleteCurrencyView(ManageCurrencyView):
         super().__init__(ctx, currency)
 
     @discord.ui.button(label="Delete", style=discord.ButtonStyle.red)
-    @Config.has_permission("manage_currencies")
     async def delete(
         self, interaction: discord.Interaction, _: discord.ui.Button
     ) -> None:
+        if self.currency.owner_id != interaction.user.id:
+            raise NotOwner("You do not own this currency")
+
         async with self._ctx.bot.pool.acquire() as con:
             await con.execute("DELETE FROM currencies WHERE id = $1;", self.currency.id)
             await con.execute(
@@ -61,7 +75,9 @@ class DeleteCurrencyView(ManageCurrencyView):
             await interaction.message.delete()
 
     @discord.ui.button(label="No", style=discord.ButtonStyle.gray)
-    @Config.has_permission("manage_currencies")
     async def no(self, interaction: discord.Interaction, _: discord.ui.Button) -> None:
+        if self.currency.owner_id != interaction.user.id:
+            raise NotOwner("You do not own this currency")
+
         if interaction.message:
             await interaction.message.delete()
